APIPython/requirements.txt
azure-storage-blob==2.1.0
click==7.1.2
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
jsonify==0.5
MarkupSafe==2.0.0
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
Werkzeug==2.0.0